email_automation/requirements.txt
psycopg2-binary>=2.9.9,<3.0
openpyxl>=3.1.2,<4.0
python-dotenv>=1.0.1,<2.0
pytest>=7.4.0,<9.0
